Seperate editor logic from the tui
