+++ src/core/mod.rs
@@ -1,5 +1,5 @@
 mod terminal;
-pub use self::terminal::{Terminal, TerminalEvent};
+pub use self::terminal::{Terminal, TerminalEvent, RenderTarget};
 
 mod tui;
 pub use self::tui::{Tui, TuiServiceBuilder, TuiService, CoreEvent};
