+++ src/widgets/mod.rs
@@ -1,3 +1,8 @@
+/// A widget is something that can be displayed on screen
+
 mod view;
 pub use self::view::View;
 pub use self::view::ViewClient;
+
+mod editor;
+pub use self::editor::Editor;
